flask-port/requirements.txt
flask==3.1.0
gunicorn==23.0.0
cryptography>=44.0.0
PyJWT>=2.10.0
httpx[http2]>=0.28.0
google-auth>=2.38.0
requests>=2.32.0
psycopg2-binary>=2.9.9
